refactor: separate replay chart from contract replay
